#  add channels last support for AvgPool2d on CPU path