--- a/aten/src/ATen/native/Pool.h
+++ b/aten/src/ATen/native/Pool.h
@@ -1,13 +1,19 @@
 #include <ATen/ATen.h>
-#include <ATen/Parallel.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/div_rtn.h>
-#include <tuple>
+#include <ATen/native/DispatchStub.h>
 
 #pragma once
 
 namespace at {
 namespace native {
+
+using max_pool2d_fn = void(*)(Tensor& output, Tensor& indices, const Tensor& input,
+    int kW, int kH, int dW, int dH, int padW, int padH, int dilationW, int dilationH);
+using max_pool2d_backward_fn = void(*)(Tensor& grad_input, const Tensor& grad_output, const Tensor& indices);
+
+DECLARE_DISPATCH(max_pool2d_fn, max_pool2d_kernel);
+DECLARE_DISPATCH(max_pool2d_backward_fn, max_pool2d_backward_kernel);
 
 namespace {