… aten/src/ATen/native/DilatedMaxPool2d.cpp
… aten/src/ATen/native/cpu/AdaptiveAvgPoolKernel.cpp
… aten/src/ATen/native/cpu/AvgPoolKernel.cpp
… aten/src/ATen/native/cpu/MaxPoolKernel.cpp
… aten/src/ATen/native/cpu/utils.h
… test/test_nn.py