This patch adds channels last support for nn.AvgPool2d for CPU.
For channels last memory format, vectorization is done on dimension of C.

On Intel(R) Xeon(R) Platinum 8260 CPU @ 2.40GHz, 2*20 cores, comparing CL v.s. contiguous:

* single core inference (BS=1): 6.7x speedup
* single socket inference (BS=1): 3.7x speedup
* single socket inference (BS=128): 2.5x speedup

```bash
(before)
### single core inference
AvgPool2d(contiguous): input_size [1, 64, 112, 112] time: 2.100 ms

### single socket inference
AvgPool2d(contiguous): input_size [1, 64, 112, 112] time: 0.132 ms
AvgPool2d(contiguous): input_size [128, 64, 112, 112] time: 16.533 ms

(after)
### single core inference
AvgPool2d(contiguous): input_size [1, 64, 112, 112] time: 2.032 ms
AvgPool2d(channels_last): input_size [1, 64, 112, 112] time: 0.304 ms

### single socket inference
AvgPool2d(contiguous): input_size [1, 64, 112, 112] time: 0.119 ms
AvgPool2d(channels_last): input_size [1, 64, 112, 112] time: 0.032 ms
AvgPool2d(contiguous): input_size [128, 64, 112, 112] time: 14.420 ms
AvgPool2d(channels_last): input_size [128, 64, 112, 112] time: 5.784 ms
```

This patch depends on #42104, #42719