@VitalyFedyunin 2020-11-16
Hi! It is hard to land and review large PR. Can you please split it by adaptive_avg avg and max pools. Stack them if necessary. Thanks.