@mingfeima 2020-11-17
> please split it by adaptive_avg avg and max pools

@VitalyFedyunin, I have requested permission to ghstack on the slack channel, I will stack the CPU channels last PRs (existing and future ones) once  I have access to ghstack.